Shutdown on exception instead of on function end
This makes it possible to use the same decorator connection.close() and
process_clipboard() to shut down the application if there is an error or
the connection closes.

It is also necessary to handle printing out the exception ourselves,
since the current thread might end due to conn.close() and root.destroy()
and before it gets the chance to print out anything.