Make server_socket reusable
Usually, linux will keep the server socket reserved for some time after
it is closed. This leads to the script failing if it is quit and then
re-run shortly after. By marking the socket as reusable, the same
address-port combo can be used immediately when the old socket is
closed.